+++ src/kernel/contexts/source/pen_context.cpp
@@ -241,6 +241,14 @@ int pen_contextReaderMat::storeString(const std::string& pathInSection,
 	mats.back().filename = element;
       return SUCCESS;
     }
+    else if(pathInSection.compare("DB") == 0){
+      mats.back().DB = element;
+      return SUCCESS;
+    }
+    else if(pathInSection.compare("DB-material") == 0){
+      mats.back().matDB = element;
+      return SUCCESS;
+    }
   }else if(family == -1){
     if(pathInSection.compare("context-log") == 0){
       contextlogfile = element;
@@ -806,10 +814,11 @@ int pen_context::configure(const double EMAX,
     if(!fileAccessible || matData.forceCreation){
 
       //Create the material
-      if(matData.composition.size() == 0){
+      if(matData.composition.size() == 0 &&
+	 matData.DB.compare("-") == 0){
 	if(verbose > 0){
 	  printf("Error: Unable to create material %s."
-		 " No composition provided.\n",
+		 " No composition nor DB provided.\n",
 		 matData.name.c_str());
 	}
 	return MISSING_MATERIAL_COMPOSITION;
@@ -827,11 +836,22 @@ int pen_context::configure(const double EMAX,
 
       //Create the material
       std::string errorString;
-      err = penred::penMaterialCreator::createMat(matData.name,
-						  matData.density,
-						  matData.composition,
-						  errorString,
-						  matData.filename);
+
+      if(matData.composition.size() > 0){
+	err = penred::penMaterialCreator::createMat(matData.name,
+						    matData.density,
+						    matData.composition,
+						    errorString,
+						    matData.filename);
+      }
+      else{
+	err = penred::penMaterialCreator::createMat(matData.name,
+						    matData.DB,
+						    matData.matDB,
+						    errorString,
+						    matData.filename);
+      }
+      
       if(err != 0){
 	if(verbose > 0){
 	  printf("Unable to create material %s.\n"
@@ -842,7 +862,7 @@ int pen_context::configure(const double EMAX,
 		 err);
 	}
 	return MATERIAL_CREATION_FAILED;
-      }
+      }      
     }
 #else
     if(!fileAccessible){